requirements.txt
Flask
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.4
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Flask-Mail==0.9.1
email-validator==1.3.1
Werkzeug==2.3.7
